Improve console logging and rendering process
This commit brings improvements in developer experience by improving and cleaning console logs and putting the outputs of unit and conformance test runs in a separate files.

Some of the logs I considered unnecessary I removed, some are moved to debug. What was added is printing of files that are sent and received from server (with line numbers and token size).

The commit adds rich library (apparently the best in python world for this) that I found very usable for what we are planning to do on client side in the future.

This commit also improves code rendering in case when unittest were not properly generated after functional requirement implementation.